requirements.txt
appdirs==1.4.2
Django==1.11b1
djangorestframework==3.5.4
packaging==16.8
pyparsing==2.1.10
pytz==2016.10
six==1.10.0
djangorestframework-csv==2.0.0
